Added unittest example
